Math thing

+++ common/q_math.h
@@ -44,7 +44,7 @@ namespace math
 
 		float Length()
 		{
-			return sqrt(x*x + y*y + z*z);
+			return sqrtf(x*x + y*y + z*z);
 		}
 
 		void Scale(float scale)
